python/requirements.txt
pyarrow
pandas
grpcio
grpcio-tools
protobuf
streamlit
plotly
